+++ src/Repository/ItemRepository.php
@@ -1,0 +1,61 @@
+<?php
+
+/**
+ * This file is part of the TodoList package.
+ *
+ * (c) Doug Harple <dharple@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Repository;
+
+use App\Entity\Item;
+use App\Entity\User;
+use DateTime;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Common\Collections\Criteria;
+use Doctrine\Persistence\ManagerRegistry;
+use Exception;
+use Psr\Log\LoggerInterface;
+
+/**
+ * Repository for Item Entity.
+ *
+ * @method Item|null find($id, $lockMode = null, $lockVersion = null)
+ * @method Item|null findOneBy(array $criteria, array $orderBy = null)
+ * @method Item[]    findAll()
+ * @method Item[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class ItemRepository extends ServiceEntityRepository
+{
+    /**
+     * Instantiates a new repository.
+     *
+     * @param ManagerRegistry $registry Mandatory.
+     */
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, Item::class);
+    }
+
+    /**
+     * Get open item count.
+     *
+     * @param User $user The user to get an open item count for.
+     *
+     * @return int
+     */
+    public function getOpenItemCount(User $user): int
+    {
+        $qb = $this->createQueryBuilder('i')
+            ->select('COUNT(i.id)')
+            ->where('i.user = :user')
+            ->andWhere('i.status = :status')
+            ->setParameter('user', $user)
+            ->setParameter('status', 'Open');
+
+        return $qb->getQuery()->getSingleScalarResult();
+    }
+}